Move `TransformRequestBodyAction` to separate file
